+++ src/main/kotlin/com/otakuexchange/infra/tables/parimutuel/MarketPoolTable.kt
@@ -12,7 +12,7 @@ object MarketPoolTable : Table("market_pools") {
     val entityId  = uuid("entity_id").references(EntityTable.id, onDelete = ReferenceOption.SET_NULL).nullable()
     val label     = text("label")
     val isWinner  = bool("is_winner").default(false)
-    val amount    = integer("amount").default(0)   // total cents staked in this pool
+    val amount    = long("amount").default(0)   // total cents staked in this pool
     val createdAt = timestamp("created_at")
     val updatedAt = timestamp("updated_at")
